Add files via upload
Interpolating a linear function from  R_13 (features)-> R_1 (label), we use ridge regression to reduce unnessecary complexity against overfitting. but we want to know how much we want to reduce this complexity by so we test on different complexity restrictions values (lambdas) and check for which of the ten cross validations the lambdas perform the best. we output the average of what we get from scores_cross_validation() for a given lambda.